ci: checkout PR head sha
